Refactor radiation-related classes and introduce new interfaces for radiation capabilities
