+++ src/main/java/net/nuclearteam/createnuclear/api/radiation/IRadiationSource.java
@@ -1,4 +1,4 @@
-package net.nuclearteam.createnuclear.content.effects;
+package net.nuclearteam.createnuclear.api.radiation;
 
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
